pass ULONG where ParseDisplayName writes one
The same mismatch as the registry and WriteFile out parameters earlier on this
branch, one layer further out: IShellFolder::ParseDisplayName takes ULONG *
for the characters it consumed and the attributes it returns, and the port had
left both locals as uint32_t. unsigned long and unsigned int are distinct
types to MSVC whatever their width.

+++ Versions/Temporary/Engine/Sources/MapEditor/PC_StringDirRefEditor.cpp
@@ -120,8 +120,8 @@ void CPCStringDirRefEditor::OnBrowse()
 			if ( SUCCEEDED( hResult ) )
 			{
 				LPITEMIDLIST pidl = NULL;
-				uint32_t dwEaten   = 0;
-				uint32_t dwAttribs = 0;
+				ULONG dwEaten   = 0;
+				ULONG dwAttribs = 0;
 
 				// no _MAX_PATH buffer to overflow, and no silent truncation of
 				// a longer path either
